Backend/requirements.txt
# Web Framework
Django==4.2.7
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter

# Face Recognition
deepface
tf-keras

# Computer Vision (phone detection via YOLOv8)
ultralytics
opencv-python-headless

# Image handling (required by Django ImageField)
Pillow

# PDF report generation
reportlab

# WebAuthn (Passkey) support
webauthn==2.1.0
